safari-600.7-branch/LayoutTests/platform/ios-sim/storage/indexeddb/objectstore-basics-workers-expected.txt

[Worker] Test the basics of IndexedDB's IDBObjectStore.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/objectstore-basics.js
[Worker] indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;
[Worker] 
[Worker] dbname = "objectstore-basics.js"
[Worker] indexedDB.deleteDatabase(dbname)
FAIL [Worker] indexedDB.deleteDatabase(dbname) threw exception TypeError: null is not an object (evaluating 'indexedDB.deleteDatabase')
FAIL [Worker] Unexpected error: TypeError: undefined is not an object (evaluating 'deleteRequest.onerror = unexpectedErrorCallback')
PASS successfullyParsed is true

TEST COMPLETE
Got error from worker: TypeError: undefined is not an object (evaluating 'deleteRequest.onerror = unexpectedErrorCallback')
PASS successfullyParsed is true

TEST COMPLETE
FAIL Unexpected error: TypeError: undefined is not an object (evaluating 'deleteRequest.onerror = unexpectedErrorCallback')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
